requirements.txt
anyio==4.3.0
asgiref==3.7.2
astroid==3.1.0
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
Flask==3.0.2
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
Logbook==1.7.0.post0
MarkupSafe==2.1.5
mccabe==0.7.0
platformdirs==4.2.0
pylint==3.1.0
python-dotenv==1.0.1
Represent==2.1
requests==2.31.0
rush==2021.4.0
sniffio==1.3.1
spacetrack==1.2.0
tomlkit==0.12.4
urllib3==2.2.1
Werkzeug==3.0.1
